Clean up icon helpers for less noisy view rendering
