+++ app/helpers/good_job/application_helper.rb
@@ -2,6 +2,11 @@
 
 module GoodJob
   module ApplicationHelper
+    # Explicit helper inclusion because ApplicationController inherits from the host app.
+    #
+    # We can't rely on +config.action_controller.include_all_helpers = true+ in the host app.
+    include IconsHelper
+
     def format_duration(sec)
       return unless sec
 
@@ -22,42 +27,6 @@ module GoodJob
       options = options.reverse_merge({ scope: "good_job.datetime.distance_in_words" })
       text = t("good_job.helpers.relative_time.#{timestamp.future? ? 'future' : 'past'}", time: time_ago_in_words(timestamp, **options))
       tag.time(text, datetime: timestamp, title: timestamp)
-    end
-
-    STATUS_ICONS = {
-      discarded: "exclamation",
-      succeeded: "check",
-      queued: "dash_circle",
-      retried: "arrow_clockwise",
-      running: "play",
-      scheduled: "clock",
-    }.freeze
-
-    STATUS_COLOR = {
-      discarded: "danger",
-      succeeded: "success",
-      queued: "secondary",
-      retried: "warning",
-      running: "primary",
-      scheduled: "secondary",
-    }.freeze
-
-    def status_badge(status)
-      content_tag :span, status_icon(status, class: "text-white") + t(status, scope: 'good_job.status', count: 1),
-                  class: "badge rounded-pill bg-#{STATUS_COLOR.fetch(status)} d-inline-flex gap-2 ps-1 pe-3 align-items-center"
-    end
-
-    def status_icon(status, **options)
-      options[:class] ||= "text-#{STATUS_COLOR.fetch(status)}"
-      icon = render_icon STATUS_ICONS.fetch(status)
-      content_tag :span, icon, **options
-    end
-
-    def render_icon(name, **options)
-      # workaround to render svg icons without all of the log messages
-      partial = lookup_context.find_template("good_job/shared/icons/#{name}", [], true)
-      options[:class] = Array(options[:class]).join(" ")
-      partial.render(self, { class: options[:class] })
     end
 
     def translate_hash(key, **options)
